requirements.txt
streamlit==1.51.0
tensorflow==2.15.0
tensorflow-intel==2.15.0
keras==2.15.0
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.25.2
pillow==12.0.0
protobuf==4.25.8
imutils==0.5.4
